* fixed some things from revision 10010 * byte fieldSize is int fieldSize again, because fieldSize might be used in the format string net read * fixed order in fb_list * fixed fb_list array size * G_MoveCalc size parameter is now fieldSize instead of 1 or 2 * G_ACTOR2x2Spawn is now G_Actor2x2Spawn * extended Grid_CheckForbidden (but this is untested and still wip)

+++ src/game/g_local.h
@@ -567,7 +567,7 @@ struct edict_s {
 	int count;		/* general purpose 'amount' variable - set via mapeditor often */
 	int sounds;		/* type of sounds to play - e.g. doors */
 	int dmg;		/* damage done by entity */
-	byte fieldSize;	/* ACTOR_SIZE_* */
+	int fieldSize;	/* ACTOR_SIZE_* */
 
 	/* function to call when used */
 	void (*use) (edict_t * self, edict_t * other, edict_t * activator);
